# [SPARK-39347][SS] Bug fix for time window calculation when event time < 0