@HeartSaVioR 2023-02-06
SPARK-38069 was merged in Spark 3.3. Porting back to master/3.4/3.3 would be enough.